% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpp.R
\name{ccmpp_migr_mid}
\alias{ccmpp_migr_mid}
\alias{ccmpp_migr_end}
\alias{demproj}
\title{Calculate a demographic projection}
\usage{
ccmpp_migr_mid(par)

ccmpp_migr_end(par)

demproj(par, proj.method = ccmpp_migr_mid, spec.fert = TRUE)
}
\arguments{
\item{par}{demographic inputs as prepared by \code{read_upd}}

\item{proj.method}{the CCMPP function to use, either \code{ccmpp_migr_mid}
for mid-period net migration or \code{ccmpp_migr_end} for end-period net
migration}

\item{spec.fert}{If \code{spec.fert=FALSE}, then proportionate age-specific
fertility \code{par$pasfrs} is used as-is. If \code{spec.fert=TRUE},
age-specific fertility is aggregated to five-year age groups, then assumed
uniform within those age groups during model projection as in Spectrum. By
default, \code{spec.fert=TRUE}.}
}
\value{
Returns a list containing
 \enumerate{
 \item{pop - an array of projected estimates by year, sex (1:male, 2:female) and age. The maximum age corresponds to the open 80+ age group.}
 \item{births - an array of births by year and child's sex.}
 \item{deaths - an array of projected deaths by year, sex, and age.}
 }
}
\description{
Calculate a demographic projection
}
\details{
This implementation is intended to replicate the demographic
  projection calculations done in the Spectrum software suite, which in turn
  is based on methods used by the United Nations Population Division to
  produce its World Population Prospects (WPP).

  The 2022 WPP revision used end period net migration. Previous WPP revisions
  used mid-period net migration. For consistency with WPP, Spectrum used
  mid-period net migration prior to Spectrum version 6.19. Since version 6.20
  released in 2022, Spectrum has used end-period net migration.

  Please note that Spectrum handles net migration inputs differently than
  DemProjR. DemProjR uses net migration inputs as specified in par$migr
  exactly. To simplify its user interface, net migration in Spectrum is
  entered by five-year age groups, then disaggregated to single ages during
  projection. This may cause small differences in population projections
  between Spectrum and DemProjR.
}
\section{Functions}{
\itemize{
\item \code{ccmpp_migr_mid()}: CCMPP that spreads net migrants uniformly throughout each period

\item \code{ccmpp_migr_end()}: CCMPP that spreads net migrants at the end of each period

}}
